Fixed FLAC 24-bit audio playback support and added the code needed to calc TotalTime for FLACs
Turns out my code for constructing the playback buffer was too crude and still has flaws to be addressed in another commit.
The problem I have left to be resolved is for 8 and 12-bit FLACs where the engine still won't downsample correctly, nor will 32-bit FLACs which are currently an impossibility.
4-bit FLACs are also an issue (upsample them?) and I have not yet decided on how to handle a few of the other supported bits-per-samples it supports